refactor(comment): rename comment repo to post comment repo + move under comment directory
