+++ lib/viewmodels/comments_view_model.dart
@@ -4,7 +4,7 @@ import "package:collection/collection.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/models/database/post/post_id_firestore.dart";
 import "package:proxima/models/ui/comment_details.dart";
-import "package:proxima/services/database/comment_repository_service.dart";
+import "package:proxima/services/database/comment/post_comment_repository_service.dart";
 import "package:proxima/services/database/user_repository_service.dart";
 
 /// This view model is used to fetch the comments of a post.
