+++ lib/viewmodels/new_comment_view_model.dart
@@ -6,7 +6,7 @@ import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/models/database/comment/comment_data.dart";
 import "package:proxima/models/database/post/post_id_firestore.dart";
 import "package:proxima/models/ui/validation/new_comment_validation.dart";
-import "package:proxima/services/database/comment_repository_service.dart";
+import "package:proxima/services/database/comment/post_comment_repository_service.dart";
 import "package:proxima/viewmodels/login_view_model.dart";
 
 /// The view model for adding a new comment to a post whose
