+++ test/mocks/data/firestore_comment.dart
@@ -2,7 +2,7 @@ import "package:proxima/models/database/comment/comment_data.dart";
 import "package:proxima/models/database/comment/comment_firestore.dart";
 import "package:proxima/models/database/comment/comment_id_firestore.dart";
 import "package:proxima/models/database/post/post_id_firestore.dart";
-import "package:proxima/services/database/comment_repository_service.dart";
+import "package:proxima/services/database/comment/post_comment_repository_service.dart";
 
 import "comment_data.dart";
 
@@ -16,7 +16,7 @@ class CommentFirestoreGenerator {
   Future<List<CommentFirestore>> addComments(
     int number,
     PostIdFirestore postId,
-    CommentRepositoryService commentRepository,
+    PostCommentRepositoryService commentRepository,
   ) async {
     final comments = <CommentFirestore>[];
