+++ test/mocks/data/firestore_user_comment.dart
@@ -3,7 +3,7 @@ import "package:proxima/models/database/comment/comment_id_firestore.dart";
 import "package:proxima/models/database/user/user_id_firestore.dart";
 import "package:proxima/models/database/user_comment/user_comment_data.dart";
 import "package:proxima/models/database/user_comment/user_comment_firestore.dart";
-import "package:proxima/services/database/user_comment_repository_service.dart";
+import "package:proxima/services/database/comment/user_comment_repository_service.dart";
 
 import "user_comment_data.dart";
