+++ test/mocks/services/mock_comment_repository_service.dart
@@ -3,10 +3,10 @@ import "package:proxima/models/database/comment/comment_data.dart";
 import "package:proxima/models/database/comment/comment_firestore.dart";
 import "package:proxima/models/database/comment/comment_id_firestore.dart";
 import "package:proxima/models/database/post/post_id_firestore.dart";
-import "package:proxima/services/database/comment_repository_service.dart";
+import "package:proxima/services/database/comment/post_comment_repository_service.dart";
 
 class MockCommentRepositoryService extends Mock
-    implements CommentRepositoryService {
+    implements PostCommentRepositoryService {
   @override
   Future<CommentIdFirestore> addComment(
     PostIdFirestore? parentPostId,
