+++ test/services/database/post_repository_test.dart
@@ -8,7 +8,7 @@ import "package:proxima/models/database/post/post_location_firestore.dart";
 import "package:proxima/models/database/user/user_id_firestore.dart";
 import "package:proxima/models/database/vote/vote_firestore.dart";
 import "package:proxima/models/database/vote/vote_state.dart";
-import "package:proxima/services/database/comment_repository_service.dart";
+import "package:proxima/services/database/comment/post_comment_repository_service.dart";
 import "package:proxima/services/database/post_repository_service.dart";
 import "package:proxima/services/database/upvote_repository_service.dart";
 
@@ -62,7 +62,8 @@ void main() {
       final dbPost = await postRef.get();
       expect(dbPost.exists, true);
 
-      final commentRepository = CommentRepositoryService(firestore: firestore);
+      final commentRepository =
+          PostCommentRepositoryService(firestore: firestore);
       await CommentFirestoreGenerator()
           .addComments(4, post.id, commentRepository);
