+++ test/viewmodels/comment_view_model_integration_test.dart
@@ -5,7 +5,7 @@ import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/models/database/comment/comment_data.dart";
 import "package:proxima/models/database/post/post_id_firestore.dart";
 import "package:proxima/models/ui/comment_details.dart";
-import "package:proxima/services/database/comment_repository_service.dart";
+import "package:proxima/services/database/comment/post_comment_repository_service.dart";
 import "package:proxima/services/database/firestore_service.dart";
 import "package:proxima/services/database/user_repository_service.dart";
 import "package:proxima/viewmodels/comments_view_model.dart";
@@ -21,7 +21,7 @@ void main() {
     late CommentDataGenerator commentDataGenerator;
 
     late UserRepositoryService userRepository;
-    late CommentRepositoryService commentRepository;
+    late PostCommentRepositoryService commentRepository;
     late AutoDisposeFamilyAsyncNotifierProvider<CommentsViewModel,
         List<CommentDetails>, PostIdFirestore> commentViewModelProvider;
     late ProviderContainer container;
