+++ test/viewmodels/new_comment_view_model_integration_test.dart
@@ -5,7 +5,7 @@ import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/models/database/post/post_id_firestore.dart";
 import "package:proxima/models/database/user/user_id_firestore.dart";
 import "package:proxima/models/ui/validation/new_comment_validation.dart";
-import "package:proxima/services/database/comment_repository_service.dart";
+import "package:proxima/services/database/comment/post_comment_repository_service.dart";
 import "package:proxima/services/database/firestore_service.dart";
 import "package:proxima/viewmodels/login_view_model.dart";
 import "package:proxima/viewmodels/new_comment_view_model.dart";
@@ -20,7 +20,7 @@ void main() {
     late FakeFirebaseFirestore fakeFirestore;
     late CommentDataGenerator commentDataGenerator;
 
-    late CommentRepositoryService commentRepository;
+    late PostCommentRepositoryService commentRepository;
     late AsyncNotifierFamilyProvider<NewCommentViewModel, NewCommentValidation,
         PostIdFirestore> newCommentViewModelPostProvider;
